+++ src/MoptAvalara6.php
@@ -17,8 +17,8 @@ use MoptAvalara6\Bootstrap\Form;
 class MoptAvalara6 extends Plugin
 {
     const PLUGIN_NAME = 'MoptAvalara6';
-
-    const PLUGIN_VERSION = '4.0.2';
+    const PLUGIN_VERSION = '4.1.0';
+    const AVALARA_APP_VERSION = 'a0n5a00000hvQLaAAM';
 
     /**
      * @param InstallContext $installContext
@@ -152,6 +152,10 @@ class MoptAvalara6 extends Plugin
                     'name' => Form::CUSTOM_FIELD_AVALARA_PRODUCT_TAX_CODE,
                     'type' => CustomFieldTypes::TEXT,
                     'allow_cart_expose' => true,
+                    'config' => [
+                        'componentName' => 'avalara-tax-code-select',
+                        'customFieldType' => 'text',
+                    ],
                 ]
             ],
             'relations' => [
@@ -183,6 +187,10 @@ class MoptAvalara6 extends Plugin
                     'name' => Form::CUSTOM_FIELD_AVALARA_CATEGORY_TAX_CODE,
                     'type' => CustomFieldTypes::TEXT,
                     'allow_cart_expose' => true,
+                    'config' => [
+                        'componentName' => 'avalara-tax-code-select',
+                        'customFieldType' => 'text',
+                    ],
                 ]
             ],
             'relations' => [
@@ -213,6 +221,10 @@ class MoptAvalara6 extends Plugin
                     'id' => Uuid::randomHex(),
                     'name' => Form::CUSTOM_FIELD_AVALARA_CUSTOMER_CODE,
                     'type' => CustomFieldTypes::TEXT,
+                    'config' => [
+                        'componentName' => 'avalara-exemption-code-select',
+                        'customFieldType' => 'text',
+                    ],
                 ]
             ],
             'relations' => [